+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unreleased
 * GooglePay
-  * Bump `play-services-wallet` version to `19.1.0`
+  * Bump `play-services-wallet` version to `19.2.1`
   * Add `totalPriceLabel` to `GooglePayRequest`
 
 ## 4.39.0 (2023-10-16)